+++ DejaVu/DejaVu_Warlock/DejaVu_Warlock.lua
@@ -1,0 +1,4 @@
+local addonName, addonTable = ... -- luacheck: ignore addonTable
+local DejaVu = _G["DejaVu"]
+local logging = DejaVu.Logging
+logging(addonName .. " loaded.")